+++ src/routes/Main/Dashboard/Client/HomePage.js
@@ -11,7 +11,9 @@ import { register,
     callResponse,
     incomingCall,
     startCommunication,
-    stop
+    stop as stopCall,
+    onReceiveFinishRequest,
+    rejectCall
 } from "../../../../classes/Connection";
 import { getConfig } from "../../../../Config";
 import { webSocketController } from "../../../../classes/WebSocket";
@@ -20,6 +22,7 @@ import { getAllAdminsOnline } from "../../../../appRedux/actions/Admin";
 import { Layout } from "../../../../components/Layout";
 import { Users } from "../../../../components/Users";
 import Icon from "../../../../components/Icon";
+import incomingSound from '../../../../assets/sounds/incoming_call.mp3';
 
 const WebRtcPeerClass = require('../../../../classes/WebRtcPeer');
 const WebRtcPeer = new WebRtcPeerClass();
@@ -31,12 +34,25 @@ export default memo(() => {
     const username = localStorage.getItem('username')
     const role = localStorage.getItem('role')
     const WEB_SOCKET_URL = getConfig('WEB_SOCKET_URL');
+    let [playAudio, setPlayAudio] = useState(false);
     
     const fetchUserData = () => {
         dispatch(getAllAdminsOnline())
             .then((res) => {
                 setUsers(res)
             })
+    }
+
+    const play = () => {
+        setPlayAudio(true);
+        playAudio = new Audio(incomingSound);
+        playAudio.loop = true;
+        playAudio.play();
+    };
+
+    const stop = () => {
+        setPlayAudio(false);
+        playAudio.pause();
     }
 
     const onSignOut = () => {
@@ -84,14 +100,13 @@ export default memo(() => {
     const navigateTo = () => history.push('/dashboard/room/call');
 
     async function connWS() {
-        // const ws = webSocketController.connect('wss://localhost:3030/one2one');
         const ws = webSocketController.connect(WEB_SOCKET_URL);
+
         ws.onopen = async function() {
             console.info('Connection with websocket server opened');
             await register(username);
         }
-        // const ws = webSocketController.connect('wss://192.168.1.101:3030/one2one');
-        // const ws = webSocketController.connect('wss://192.168.1.100:3030/one2one');
+
         ws.onmessage = async function(message) {
             var parsedMessage = JSON.parse(message.data);
             console.info('Received message: ' + message.data);
@@ -104,27 +119,48 @@ export default memo(() => {
                 callResponse(parsedMessage);
                 break;
             case 'incomingCall':
-                incomingCall(parsedMessage);
-                localStorage.setItem('me', username)
+                play();
+                Modal.confirm({
+                    title: "Panggilan Masuk",
+                    icon: <Icon name="alert" width={24} height={24} />,
+                    content:
+                        "Anda mendapatkan panggilan dari " + parsedMessage.from + ". Apakah Anda ingin menerima panggilan?",
+                    centered: true,
+                    width: 320,
+                    okText: "Terima",
+                    cancelText: "Tolak",
+                    onOk: () => {
+                        stop();
+                        incomingCall(parsedMessage);
+                        localStorage.setItem('me', username)
+                    },
+                    onCancel: () => {
+                        stop();
+                        rejectCall(parsedMessage);
+                    }
+                });
                 break;
             case 'startCommunication':
                 startCommunication(parsedMessage);
                 break;
             case 'stopCommunication':
                 console.info("Communication ended by remote peer");
-                stop(true);
+                stopCall(true);
                 const link = role === "admin" ? "/dashboard/admin/home" : "/dashboard/client/home";
                 // history.push(link);
                 localStorage.removeItem('me')
                 localStorage.removeItem('they')
                 window.location.replace(link);
                 break;
+            case 'onReceiveFinishRequest':
+                onReceiveFinishRequest(parsedMessage);
+            break;
             case 'iceCandidate':
                 const webRtcPeer = WebRtcPeer.getPeers()
                 console.log(webRtcPeer)
                 console.log('TESSSSTTTTTTT')
                 console.log(parsedMessage.candidate)
-                await webRtcPeer.addIceCandidate(parsedMessage.candidate)
+                await webRtcPeer.addIceCandidate(new RTCIceCandidate(parsedMessage.candidate))
                 // add delay 2s
                 setTimeout(() => {
                     // setLoading(false);
